+++ src/Creational/Factories/Static/SmallBox.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Codenip\Creational\Factories\Static;
+
+class SmallBox implements Box
+{
+    public function size(): string
+    {
+        return 'Small';
+    }
+}