+++ Changelog.md
@@ -1,5 +1,8 @@
 # Change Log
 
+### 2024-04-11
+- Create example of Static Factory pattern and its tests
+
 ### 2024-04-04
 - Create example of Adapter pattern and its tests